pyproject.toml
[project]
name = "lerobotsim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "genesis-world>=0.3.8",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
